+++ crates/interledger-http/src/lib.rs
@@ -15,6 +15,7 @@ mod server;
 
 // So that settlement engines can use errors
 pub mod error;
+pub mod idempotency;
 
 pub use self::client::HttpClientService;
 pub use self::server::HttpServer;
