+++ crates/interledger-settlement/src/test_helpers.rs
@@ -2,10 +2,13 @@ use super::*;
 use crate::api::scale_with_precision_loss;
 use crate::Convert;
 use crate::{LeftoversStore, SettlementEngineDetails};
+use bytes::Bytes;
 use futures::{
     future::{err, ok},
     Future,
 };
+use hyper::StatusCode;
+use interledger_http::idempotency::*;
 use interledger_service::{
     incoming_service_fn, outgoing_service_fn, Account, AccountStore, IncomingService, Username,
 };
