+++ crates/interledger-store-redis/tests/settlement_test.rs
@@ -5,8 +5,9 @@ use common::*;
 use futures::future::join_all;
 use http::StatusCode;
 use interledger_api::NodeStore;
+use interledger_http::idempotency::IdempotentStore;
 use interledger_service::{Account, AccountStore};
-use interledger_settlement::{IdempotentStore, LeftoversStore, SettlementAccount, SettlementStore};
+use interledger_settlement::{LeftoversStore, SettlementAccount, SettlementStore};
 use interledger_store_redis::AccountId;
 use lazy_static::lazy_static;
 use num_bigint::BigUint;
